trying to fix remotemediator

+++ app/src/main/java/com/example/rickandmorty/data/local/paging/remotemediator/EpisodeRemoteMediator.kt
@@ -57,7 +57,7 @@ class EpisodeRemoteMediator(val service: RickAndMortyApi, val db: AppDataBase) :
                 }
                 episodes?.forEach {
                     it.page = loadKey
-                    keyDao.insertOrReplace(PageKey(it.id.toLong(), pageInfo?.next))
+                    keyDao.insertOrReplace(PageKey(it.id, pageInfo?.next))
                 }
                 episodes?.let { episodeDao.insertAll(it) }
             }
